Moving configs and adding doc in readme
